FIX(aac3): Changed 'int64_t' to 'long long', redefinition of function
GitOrigin-RevId: fbcc49ad2d3767f87fecd5efe85f5c93d765c6e0

+++ dtests/node_data/clients/cpp/src/common/logger/Logger.cpp
@@ -92,7 +92,7 @@ Logger &Logger::operator<<(char c)
     return *this;
 }
 
-Logger &Logger::operator<<(int64_t val)
+Logger &Logger::operator<<(long long val)
 {
 
     if (level >= logConfiguration.minimumLevel) {
